Security, Authorization, Authentication update

+++ src/main/java/com/example/bookingnl/bussines/impl/UserServiceImpl.java
@@ -4,14 +4,41 @@ import com.example.bookingnl.bussines.UserService;
 import com.example.bookingnl.domain.User;
 import com.example.bookingnl.persistence.UserRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
-public class UserServiceImpl implements UserService {
+@Transactional
+@Slf4j
+public class UserServiceImpl implements UserService, UserDetailsService {
     private UserRepository repository;
+    private PasswordEncoder passwordEncoder;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Optional<User> user = repository.findByEmail(username);
+        if (user.isEmpty()) {
+            log.error("User not found");
+            throw new UsernameNotFoundException("Username not found in db");
+        } else {
+            log.info("User with email {} found", username);
+        }
+        Collection<SimpleGrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(new SimpleGrantedAuthority(user.get().getRole()));
+        return new org.springframework.security.core.userdetails.User(user.get().getEmail(), user.get().getPassword(), authorities);
+    }
 
     @Override
     public User findByEmailAndPassword(String email, String password) {
@@ -20,16 +47,33 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteByEmail(String email) {
-        repository.deleteByEmail(email);
+        Optional<User> user = repository.findByEmail(email);
+        if (user.isEmpty()) {
+            log.error("User not found");
+            throw new UsernameNotFoundException("Username not found in db");
+        } else {
+            repository.deleteByEmail(email);
+            log.info("User with email {} found", email);
+        }
     }
 
     @Override
-    public User save(User user) {
-        return repository.save(user);
+    public User save(User newUser) {
+        Optional<User> user = repository.findByEmail(newUser.getEmail());
+        if (user.isPresent()) {
+            log.error("User not found");
+            //Email taken //TODO
+            throw new UsernameNotFoundException("Username not found in db");
+        } else {
+            newUser.setPassword(passwordEncoder.encode(newUser.getPassword()));
+            return  repository.save(newUser);
+        }
     }
 
     @Override
-    public List<User> findAll (){
-       return repository.findAll();
+    public List<User> findAll() {
+        return repository.findAll();
     }
+
+
 }
